# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SkyHigh Core - Digital Check-In System API
  description: |
    API for airline self-check-in system handling seat selection, 
    baggage validation, payment processing, and boarding pass generation.
    
    ## Authentication
    All endpoints except `/api/v1/auth/*` require JWT Bearer token authentication.
    
    ## Rate Limiting
    - Login: 5 requests per 15 minutes per booking reference
    - Seat map: 60 requests per minute
    - Other endpoints: 30 requests per minute
  version: 1.0.0
  contact:
    name: SkyHigh Airlines Support
    email: support@skyhigh.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/seats/{seatId}/hold:
    post:
      tags:
        - Seats
      summary: Hold a seat
      description: Reserve a seat for 120 seconds
      parameters:
        - name: seatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HoldSeatRequest'
      responses:
        '200':
          description: Seat held successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatHoldResponse'
        '409':
          description: Seat already held or confirmed

    delete:
      tags:
        - Seats
      summary: Release seat hold
      parameters:
        - name: seatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Seat released
        '422':
          description: Cannot release seat
...
  /api/v1/check-in/start:
    post:
      tags:
        - Check-In
      summary: Start check-in
      description: Start a new check-in session (expires after 10 minutes)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartCheckInRequest'
      responses:
        '201':
          description: Check-in session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResponse'
        '409':
          description: Active check-in already exists
        '422':
          description: Check-in window not open
...
  /api/v1/check-in/{checkInId}/payment:
    post:
      tags:
        - Check-In
      summary: Process payment
      description: Process payment for excess baggage fee
      parameters:
        - name: checkInId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResponse'
        '422':
          description: Payment failed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PassengerInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    FlightInfo:
      type: object
      properties:
        flightId:
          type: string
          format: uuid
        flightNumber:
          type: string
        departureTime:
          type: string
          format: date-time
        origin:
          type: string
        destination:
          type: string
        checkInOpen:
          type: boolean
        checkInOpensAt:
          type: string
          format: date-time
        checkInClosesAt:
          type: string
          format: date-time
...
    HoldSeatRequest:
      type: object
      required:
        - checkInId
      properties:
        checkInId:
          type: string
          format: uuid
    SeatHoldResponse:
      type: object
      properties:
        seatId:
          type: string
          format: uuid
        seatNumber:
          type: string
        seatClass:
          type: string
        status:
          type: string
        heldUntil:
          type: string
          format: date-time
        holdDurationSeconds:
          type: integer
    StartCheckInRequest:
      type: object
      required:
        - flightId
        - bookingReference
      properties:
        flightId:
          type: string
          format: uuid
        bookingReference:
          type: string
...
    PaymentRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
          default: INR
        paymentMethod:
          type: string
          default: CARD
        idempotencyKey:
          type: string
    CheckInResponse:
      type: object
      properties:
        checkInId:
          type: string
          format: uuid
        status:
          type: string
          enum: [IN_PROGRESS, WAITING_PAYMENT, COMPLETED, CANCELLED, EXPIRED]
        flight:
          $ref: '#/components/schemas/FlightInfo'
        passenger:
          $ref: '#/components/schemas/PassengerInfo'
        seat:
          type: object
          properties:
            seatId:
              type: string
              format: uuid
            seatNumber:
              type: string
            seatClass:
              type: string
            heldUntil:
              type: string
              format: date-time
        baggage:
          type: object
          properties:
            weightKg:
              type: number
            maxAllowedKg:
              type: number
            excessKg:
              type: number
            excessFee:
              type: number
            currency:
              type: string
            paymentRequired:
              type: boolean
        startedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        nextSteps:
          type: array
          items:
            type: string
        boardingPass:
          type: object
          properties:
            id:
              type: string
              format: uuid
            barcode:
              type: string
            downloadUrl:
              type: string
...